refactor(dsh-notifier): PR3 注释清理与失真修正 + locales 死键 + 客户端遗留修复（#669）
- src 全量清除外部 issue 号与会话/阶段/决策/缺陷内部编号，保留 why 与领域不变量名
- 两处残留失真注释改写（NotifyRequest.body 脱敏口径、message.ts 文案单表）
- locales 13 个零引用键 zh/en 双删（编译期锁平衡 173/173）
- 客户端 C3-2（discardChanges 关闭 409 横幅）/C3-7（冲突恢复失败专键 conflictReloadFail）/stale 文案口径（D20）/playChime 死函数清理
- 结构修复：BUILTIN_CHANNELS 物理归属下沉 config 域，消除 sdk ⇄ pipeline 值依赖环
- 类型面修复：sanitizeNoticeContent 的 enabled 容错契约（boolean | undefined）
