+++ packages/dsh-notifier/src/config/interface.ts
@@ -11,6 +11,7 @@ export {
   ASSEMBLY_SETTING_KEYS,
   BARK_ID_PATTERN,
   BARK_RESERVED_KEYS,
+  BUILTIN_CHANNELS,
   CONFIG_KEYS,
   DEFAULT_CONFIG,
   SOUND_IDS,
