+++ packages/dsh-notifier/src/events/agent-session.ts
@@ -3,9 +3,9 @@
  *
  * 事件层从 Agent 对象读取会话标题/turn 终态、判定子代理归属的最小结构子集：
  * 全部经官方类型层 import type（不引入运行时依赖），payload 跨宿主边界不受信，
- * 读取一律防御性收窄。sessionTitleOf 是全部 taskTitle 的唯一来源——B-1 后
- * 脱敏统一到 sendKind 渲染后单点（此处仅截断），taskTitle 经模板拼接与历史
- * 落盘仍全覆盖（issue #30 由统一时点承接）。
+ * 读取一律防御性收窄。sessionTitleOf 是全部 taskTitle 的唯一来源——脱敏
+ * 统一到 sendKind 渲染后单点（此处仅截断），taskTitle 经模板拼接与历史
+ * 落盘仍全覆盖（由统一时点承接）。
  */
 import type { Agent } from "@deepseek-ai/dsh-agent";
 import type { TurnEndReason } from "@deepseek-ai/dsh-session/types";
@@ -19,9 +19,9 @@ type TurnEndKind = TurnEndReason extends { kind: infer K } ? K : never;
  * 数据源：agent.session.snapshotEvents() 中最后一个 session/title 事件（dsh-session-title
  * 官方插件维护，与 GUI 会话列表同源）。无标题（新会话/未生成）返回 undefined。
  * 仅截断 40 字符（展示语义：模板拼接/taskTitle 的标题行上限）；**不再脱敏**——
- * B-1 统一时点后标题敏感片段由 sendKind 渲染后统一处理（本函数是全部 taskTitle
+ * 统一时点后标题敏感片段由 sendKind 渲染后统一处理（本函数是全部 taskTitle
  * 的唯一来源，脱敏统一到下游单点，避免此处截断先于打码造成「长敏感串被腰斩成
- * 不满足规则阈值的残段漏网」的旧问题面；issue #30 的覆盖由 sendKind 承接）。
+ * 不满足规则阈值的残段漏网」的旧问题面；该覆盖由 sendKind 承接）。
  * @param agent Agent 对象（事件 payload.agent）。
  * @returns 截断 40 字符的标题（未脱敏；sanitizeContent=false 时原样透传）。
  */
@@ -69,8 +69,7 @@ export function lastTurnEndOf(agent: Agent | undefined): { turn: number; kind: T
       if (reason === undefined || reason === null || typeof reason !== "object") continue;
       // 非有限 turn 的证据不可用（payload 跨宿主边界，不受信）：跳过该条
       // 继续向前扫——若返回 {turn:NaN} 会被完成判定当作合法证据推进
-      // lastEndedTurn=NaN，此后真实完成因 x > NaN 恒 false 被永久吞掉
-      // （#284 复核闸 P1）。
+      // lastEndedTurn=NaN，此后真实完成因 x > NaN 恒 false 被永久吞掉。
       if (typeof ev.data?.turn !== "number" || !Number.isFinite(ev.data.turn)) continue;
       const kind = String((reason as { kind?: unknown }).kind ?? "") as TurnEndKind;
       return { turn: ev.data.turn, kind };
@@ -82,7 +81,7 @@ export function lastTurnEndOf(agent: Agent | undefined): { turn: number; kind: T
 }
 
 /**
- * 运行时归属查询面（issue #49）：对齐宿主 ctx.agents（AgentRegistry）判定所需
+ * 运行时归属查询面：对齐宿主 ctx.agents（AgentRegistry）判定所需
  * 的最小结构子集——get 查活体父 agent 存在性、isOwnedBy 断言「该子 agent 确由
  * 该父 agent 的作用域创建」。仅 import type 官方 Agent，不引入运行时依赖。
  */
@@ -108,7 +107,7 @@ export interface SubagentOwnership {
  * 为何需要信号二（运行时归属）：fork 型委派与用户 fork 主线在持久化 header 上
  * **不可区分**（两者都只落 parentSession + seedLength、无 origin）——单看 header
  * 要么漏报 fork 委派（现状 bug：委派 worker 完成被误报主任务 kind=done），要么
- * 把用户 fork 主线误静默（#7 的镜像回归）。唯一可靠区分点是**运行时归属**：
+ * 把用户 fork 主线误静默（镜像回归）。唯一可靠区分点是**运行时归属**：
  * ctx.agents.get(parentSession) 非 undefined 且 isOwnedBy(该 agent id, 父 agent)
  * === true 时，该 fork 会话确由父 agent 作用域创建（委派 worker）；归属不成立
  * （父 id 不在 live registry / isOwnedBy false / agents 服务不可用）则视为用户
@@ -117,9 +116,9 @@ export interface SubagentOwnership {
  * 已知保守边界（冷 resume）：实例重启后脱离父作用域续跑的 fork worker 不在
  * live registry，归属不成立 → 退报 done。与宿主 apiproxy 冷路径 fence 行为一致
  * （attached/inspect 检查点 agent 参数为 void 0 时仅 origin 生效），属已确认的
- * 接受边界（issue #49 2026-08-22 评论源码级确认），不做持久化推断补齐。
+ * 接受边界（已源码级确认），不做持久化推断补齐。
  *
- * 运行时同型变体（issue #199 P2-1）：父 agent 先亡/被清理（如委派方提前结束、
+ * 运行时同型变体：父 agent 先亡/被清理（如委派方提前结束、
  * registry 逐出）时 `get(parent)` 同样返回 undefined，归属不成立 → 同样退报
  * done。与冷 resume 共享同一保守语义的方向性代价：宁可多报一条主任务 kind=done，
  * 不静默任何真实完成；两变体均不做持久化推断补齐。
