+++ packages/dsh-notifier/src/server/routes.ts
@@ -26,7 +26,7 @@ export const ROUTES = {
 
 /**
  * buildRoutes 的依赖注入面（index.ts 装配）。配置域部分 = ConfigPort 契约——
- * L8-2 起 setConfirm 移除，kind 确认写一律走 confirmKind（CAS 重试 ≤2）。
+ * setConfirm 已移除，kind 确认写一律走 confirmKind（CAS 重试 ≤2）。
  */
 export interface RouteDeps extends ConfigPort {
   /** 日志出口。 */
@@ -52,13 +52,13 @@ export type PatchResult =
 
 /**
  * 配置保存纯函数（PUT /config 的主体，独立导出供 smoke 单测）：
- * channels[].deviceKey 掩码按 id 对齐回填 user 层原值（评审 P0-2：必须先于
+ * channels[].deviceKey 掩码按 id 对齐回填 user 层原值（必须先于
  * 校验，掩码不是合法 key 语义）→ validateSettings 定位首个非法键（400 + hint）
- * → sanitizePatchSettings 净化（已知键校验 + 未知键透传保留、装配键剔除，
- * #470）→ 经 settings 服务 update 增量写入（expectedRevision
- * 可选做乐观并发）。错误映射（R3）：SETTINGS_CONFLICT → 409 固定文案；settings
- * 缺失 → 503 settings-unavailable；写入异常原文只进服务端日志（P2-2）。
- * 成功响应的 user 经 redactConfigView 统一脱敏（评审 P0-1 单一出口）。
+ * → sanitizePatchSettings 净化（已知键校验 + 未知键透传保留、装配键剔除）
+ * → 经 settings 服务 update 增量写入（expectedRevision
+ * 可选做乐观并发）。错误映射：SETTINGS_CONFLICT → 409 固定文案；settings
+ * 缺失 → 503 settings-unavailable；写入异常原文只进服务端日志。
+ * 成功响应的 user 经 redactConfigView 统一脱敏（脱敏单一出口）。
  */
 export async function applyConfigPatch(deps: RouteDeps, payload: unknown): Promise<PatchResult> {
   if (!deps.writable()) {
@@ -68,7 +68,7 @@ export async function applyConfigPatch(deps: RouteDeps, payload: unknown): Promi
     patch?: unknown;
     expectedRevision?: unknown;
   };
-  // D19（L8-5）：expectedRevision 仅接受「省略/null → undefined 透传」或「非负
+  // expectedRevision 仅接受「省略/null → undefined 透传」或「非负
   // 整数」；其余形态（字符串/小数/负数）显式 400，不再静默忽略。
   if (
     body.expectedRevision !== undefined &&
@@ -82,7 +82,7 @@ export async function applyConfigPatch(deps: RouteDeps, payload: unknown): Promi
       ? body.expectedRevision
       : undefined;
   const rawPatch = body.patch;
-  // 掩码回填先于校验（评审 P0-2）：channels 实例的 deviceKey 整值等于掩码时按
+  // 掩码回填先于校验：channels 实例的 deviceKey 整值等于掩码时按
   // id 对齐回填 user 层原值；新实例（user 层无同 id）带掩码 → 400 拒绝。
   let effectivePatch: unknown = rawPatch;
   if (typeof rawPatch === "object" && rawPatch !== null && Array.isArray((rawPatch as Record<string, unknown>).channels)) {
@@ -98,12 +98,12 @@ export async function applyConfigPatch(deps: RouteDeps, payload: unknown): Promi
     }
     effectivePatch = { ...(rawPatch as Record<string, unknown>), channels: unmasked.channels };
   }
-  // 先定位首个非法键（H6）：错误文案指明字段与合法范围，不再静默丢弃回默认
+  // 先定位首个非法键：错误文案指明字段与合法范围，不再静默丢弃回默认
   const invalid = validateSettings(effectivePatch);
   if (invalid !== null) {
     return { ok: false, status: 400, code: "invalid", response: { error: `配置校验失败: ${invalid.key}`, hint: invalid.hint } };
   }
-  // #470 双通道拆分后 PUT 走透传净化（未知键保留、装配键剔除）：
+  // 双通道拆分后 PUT 走透传净化（未知键保留、装配键剔除）：
   // 「至少一个有效配置键」判据改为「原始 patch 非空对象」——纯未知键 patch
   // （如 {futureKey:1}）→ 200 透传写入；仅空 patch {}（无任何键可写）→ 400。
   if (typeof rawPatch !== "object" || rawPatch === null || Object.keys(rawPatch).length === 0) {
@@ -122,12 +122,12 @@ export async function applyConfigPatch(deps: RouteDeps, payload: unknown): Promi
     if (code === "SETTINGS_CONFLICT") {
       return { ok: false, status: 409, code: "conflict", response: { error: "版本冲突", code: "SETTINGS_CONFLICT" } };
     }
-    // P2-2：对外收敛固定文案，不把底层异常原文（可能含路径等内部信息）回给
+    // 对外收敛固定文案，不把底层异常原文（可能含路径等内部信息）回给
     // 客户端；完整原因走服务端日志。
     deps.logger.warn(`dsh-notifier: 配置保存写入设置存储失败 — ${errorMessage(err)}`);
     return { ok: false, status: 500, code: "error", response: { error: "保存失败，请查看服务端日志" } };
   }
-  // 单一脱敏出口（评审 P0-1）：readUser 的 user 含 channels 明文，掩码后返回
+  // 单一脱敏出口：readUser 的 user 含 channels 明文，掩码后返回
   const fresh = deps.readUser();
   return { ok: true, value: { user: redactConfigView(fresh.user), revision: fresh.revision } };
 }
@@ -146,7 +146,7 @@ export function buildRoutes(deps: RouteDeps): WebRoute[] {
       if (!guardLoopbackMethod(req, res, ["GET", "PUT"])) return;
       if (req.method === "GET") {
         const { user, revision } = readUser();
-        // 凭据脱敏单一出口（评审 P0-1）：user 与 effective 双视图都经
+        // 凭据脱敏单一出口：user 与 effective 双视图都经
         // redactConfigView 掩码 deviceKey——深度扫描契约测试锁死两出口。
         writeJson(res, 200, {
           ok: true,
@@ -173,7 +173,7 @@ export function buildRoutes(deps: RouteDeps): WebRoute[] {
           logger.warn(`dsh-notifier: 配置请求体读取失败: ${message}`);
           return;
         }
-        // #470 复核 P0-2 兜底：applyConfigPatch 内部异常（理论不可达——净化/
+        // applyConfigPatch 内部异常兜底（理论不可达——净化/
         // 校验/脱敏对任意输入均收敛为错误分支，此处防未来改动引入未捕获抛错）
         // 不得让 handler 冒泡成宿主 500/悬挂——收敛 500 固定文案 + 服务端日志
         let result: PatchResult;
@@ -219,8 +219,8 @@ export function buildRoutes(deps: RouteDeps): WebRoute[] {
         "cache-control": "no-cache",
         connection: "keep-alive",
       });
-      // 预存在缺口随手修（#334 评审遗留 P2-9）：connected 锚点写失败（对端已断）
-      // 直接返回、不再 register——已断连接入表只会成为靠心跳/广播兜底清理的残留。
+      // connected 锚点写失败（对端已断）直接返回、不再 register——已断连接入表
+      // 只会成为靠心跳/广播兜底清理的残留。
       try {
         res.write(": connected\n\n");
       } catch {
@@ -242,11 +242,11 @@ export function buildRoutes(deps: RouteDeps): WebRoute[] {
 
   /** 健康检查：插件是否加载、配置摘要、SSE 连接数与回收观测。
    *  sseConnections 语义 = 服务端未释放的 SSE 句柄数，非「在线设备数」；上限见
-   *  config.maxConnections。#515 起连接表由 shared/sse-hub 管理：stalled 超窗 /
+   *  config.maxConnections。连接表由 shared/sse-hub 管理：stalled 超窗 /
    *  maxAge 轮换主动回收 + 上限淘汰，sseEvicts 暴露各回收路径计数（观测残留构成），
    *  sseConnHealth 暴露逐连接 age/lastWriteAgo/stalled（先量化再调参）。
-   *  #640/#641：platform（宿主平台——客户端系统卡按它显示平台提示，防浏览器 OS
-   *  与宿主 OS 混淆）+ 摘要键表补 browserSound/systemSound（D1，四同步）。 */
+   *  platform（宿主平台——客户端系统卡按它显示平台提示，防浏览器 OS
+   *  与宿主 OS 混淆）+ 摘要键表补 browserSound/systemSound（四同步）。 */
   const healthRoute: WebRoute = {
     kind: "exact",
     path: ROUTES.health,
@@ -282,7 +282,7 @@ export function buildRoutes(deps: RouteDeps): WebRoute[] {
 
   /**
    * 测试通知：POST 触发一条测试通知（绕过免打扰，测试意图是验证通道本身）。
-   * M2 收敛到 service 管线（sendKind('test')）：内置 browser/system 与配置驱动
+   * 收敛到 service 管线（sendKind('test')）：内置 browser/system 与配置驱动
    * 频道（bark）走同一分发路径——测试才有意义（验证真实投递链路），历史落盘
    * 与终态上报（status/sent 事件）同源。body 可选 {channelId}：指定单频道测试
    * （设置页频道卡「测试」按钮）。固定文案模板，不引入自由文本面（评审定案）。
@@ -333,7 +333,7 @@ export function buildRoutes(deps: RouteDeps): WebRoute[] {
   /**
    * 动态 kind 清单与确认：GET 返回注册表（含确认态，确认态持久化在配置
    * allowKinds）；POST {kind, confirmed} 写确认（仅注册表内已注册的动态 kind）。
-   * 确认动作只发生在用户主动打开设置页时（终稿 §5.1：注册即弹窗打扰不允许）。
+   * 确认动作只发生在用户主动打开设置页时（注册即弹窗打扰不允许）。
    */
   const kindsRoute: WebRoute = {
     kind: "exact",
@@ -367,8 +367,8 @@ export function buildRoutes(deps: RouteDeps): WebRoute[] {
           writeJson(res, 404, { ok: false, error: { code: "not-found", details: `未注册的动态 kind: ${kind}` } });
           return;
         }
-        // #405 PR3：confirmKind 现为 CAS 循环（可抛 SETTINGS_CONFLICT 耗尽 /
-        // 服务缺失 rejection）——handler 必须兜底（评审 P1-1），防 rejection 冒泡成
+        // confirmKind 现为 CAS 循环（可抛 SETTINGS_CONFLICT 耗尽 /
+        // 服务缺失 rejection）——handler 必须兜底，防 rejection 冒泡成
         // 宿主行为未定义；200 响应体带新 revision（向后兼容新增字段）供客户端
         // confirmOne 同步 meta——修「确认 kind 后同窗口保存必 409」的版本链断点。
         try {
@@ -380,7 +380,7 @@ export function buildRoutes(deps: RouteDeps): WebRoute[] {
             return;
           }
           if (code === "SETTINGS_UNAVAILABLE") {
-            // 与 PUT /config 的服务缺失语义一致（#405 PR3 复核）：settings 服务未
+            // 与 PUT /config 的服务缺失语义一致：settings 服务未
             // attach → 503，而非笼统 500——保持跨通道错误映射一致。
             writeJson(res, 503, { ok: false, error: { code: "settings-unavailable", error: "设置服务不可用" } });
             return;
@@ -409,7 +409,7 @@ export function buildRoutes(deps: RouteDeps): WebRoute[] {
         return;
       }
       if (req.method === "DELETE") {
-        // L8-6：clear 失败（删除/写回异常向上抛）收敛 500 固定文案、原因只进
+        // clear 失败（删除/写回异常向上抛）收敛 500 固定文案、原因只进
         // 服务端日志（对照 PUT /config 500 语义），不再恒 200。
         try {
           const removed = await history.clear();
