+++ packages/dsh-notifier/src/server/system-notifier.ts
@@ -37,9 +37,9 @@ let lastSystemOutcome = true;
  * @param options.toastScript toast.ps1 路径。
  * @param options.warn 日志出口（ctx.logger.warn）。
  * @param options.execFileImpl 探测实现注入（缺省 node:child_process execFile）——
- *   测试注入 fake 消除真实 execFile（PR0 红测先行 1；行为无关，缺省语义不变）。
+ *   测试注入 fake 消除真实 execFile（行为无关，缺省语义不变）。
  * @param options.spawnImpl spawn 实现注入（缺省 node:child_process spawn）——同上，
- *   测试注入 fake 消除真实子进程 spawn 面（T3-3）。
+ *   测试注入 fake 消除真实子进程 spawn 面。
  * @param options.killTimeoutMs 子进程超时杀进程毫秒（缺省 8000；测试注入短值）。
  */
 export function createSystemNotifier(options: {
@@ -67,9 +67,9 @@ export function createSystemNotifier(options: {
     });
   }
 
-  // #640 Linux 自播播放器探测（pw-play=PipeWire → paplay=PulseAudio；与
+  // Linux 自播播放器探测（pw-play=PipeWire → paplay=PulseAudio；与
   // notifySendAvailable 完全独立的可用性位——自播失败绝不误置 notify-send
-  // 不可用，见 error handler；P1-2 修订）。异步只探一次。
+  // 不可用，见 error handler）。异步只探一次。
   let selfPlayBin: string | undefined = undefined;
   if (process.platform === "linux") {
     exec("pw-play", ["--version"], { timeout: 3000 }, (err1) => {
@@ -94,21 +94,21 @@ export function createSystemNotifier(options: {
       try {
         // 关键：原生二进制缺失/不可执行（ENOENT 等）必须被下方 error 事件接住，
         // 绝不能冒泡成 unhandled 'error' 把宿主进程打挂——历史版本在 macOS 上因
-        // 直接 spawn powershell 失败且未挂 error 监听而崩溃（见 issue #1）。
+        // 直接 spawn powershell 失败且未挂 error 监听而崩溃。
         child = runSpawn(bin, argv, process.platform === "win32" ? { windowsHide: true, stdio: ["ignore", "ignore", "pipe"] } : { stdio: "ignore" });
       } catch (error) {
         warn(`dsh-notifier: 命令启动失败（${bin}）: ${errorMessage(error)}`);
         return resolveResult(false);
       }
       // Windows 分支限长收集 stderr：PS 的诊断（param 绑定失败、WinRT 异常等）
-      // 原先随 stdio ignore 全部丢弃，排查只能盲猜（issue #238）；只留尾部片段进日志。
+      // 原先随 stdio ignore 全部丢弃，排查只能盲猜；只留尾部片段进日志。
       let stderrTail = "";
       if (child.stderr) {
         child.stderr.on("data", (chunk: Buffer) => {
           if (stderrTail.length < 512) stderrTail += chunk.toString("utf8");
         });
       }
-      // 子进程超时兜底（8s：音频/通知进程均为短命任务；对抗评审 P2 建议 30s 过长）
+      // 子进程超时兜底（8s：音频/通知进程均为短命任务；30s 过长）
       const killer = setTimeout(() => {
         try {
           child!.kill();
@@ -142,11 +142,11 @@ export function createSystemNotifier(options: {
   /**
    * 平台 × 声音 × 弹窗的自播判定（与 text 域命令构造同域）：
    * - linux：任何非静音（true/SoundId）都自播（DE 对 hint 支持参差，toast 发声
-   *   不可依赖，#640 核心）；弹窗开也一样（suppress-sound 防双响）。
+   *   不可依赖）；弹窗开也一样（suppress-sound 防双响）。
    * - darwin：只响不弹（pop=false）自播 afplay；弹窗开 true 由 osascript 原生
    *   sound（不自播），弹窗开 SoundId 也是原生 sound（NSSound 名，不自播）。
    * - win32：只响不弹自播 SoundPlayer；弹窗开 true 由 toast 默认音（不自播，
-   *   P0-2：SoundPlayer 仅用于「应用必须自播」= 只响不弹或 SoundId 场景——
+   *   SoundPlayer 仅用于「应用必须自播」= 只响不弹或 SoundId 场景——
    *   弹窗开 SoundId 也自播（toast silent 防双响））。
    */
   function shouldSelfPlay(pop: boolean, tone: SoundSetting, platform: string): boolean {
@@ -163,15 +163,15 @@ export function createSystemNotifier(options: {
    *    spawn 失败只 warn 不翻转终态——旧契约「系统通知失败静默、仅日志、不
    *    影响主流程」：无桌面会话/无 notify-send 是常态环境而非投递失败）；
    * 2. 自播（shouldSelfPlay 判定）：Linux/Windows/macOS 音色或只响不弹场景
-   *    spawn 播放命令（同一节流窗口内——1s 节流覆盖单次投递全部 spawn，B3）。
+   *    spawn 播放命令（同一节流窗口内——1s 节流覆盖单次投递全部 spawn）。
    * 自播失败在弹窗场景只影响声音（toast 已成功，尽力而为）；「只响不弹」时
-   * 自播失败即整体失败（异步终态 failed 上报，B4/P1-2）。
+   * 自播失败即整体失败（异步终态 failed 上报）。
    * @returns 投递终态 boolean（只响不弹 = 自播成败；弹窗场景恒 true——toast
    *   成败静默，见上）。
    */
   async function deliverOnce(pop: boolean, tone: SoundSetting, rawTitle: string, rawMessage: string): Promise<boolean> {
     // 截断按码点而非 UTF-16 code unit：防 emoji 等代理对在边界被腰斩成
-    // 孤立代理（经 JSON/base64 后变成 U+FFFD 替换符显示，issue #238 配套）。
+    // 孤立代理（经 JSON/base64 后变成 U+FFFD 替换符显示）。
     const truncateCodePoints = (s: string, max: number): string => {
       const chars = Array.from(s);
       return chars.length > max ? chars.slice(0, max).join("") : s;
@@ -203,7 +203,7 @@ export function createSystemNotifier(options: {
       const snd = buildSoundCommand(platform, toneId, player);
       if (snd === null) {
         // 无播放器 / 事件文件缺失：弹窗场景声音尽力而为（忽略）；只响不弹场景
-        // 声音是唯一动作 → 失败诚实上报（B4/P1-2：不做「静默成功」）
+        // 声音是唯一动作 → 失败诚实上报（不做「静默成功」）
         if (pop === false) soundOk = false;
       } else {
         const ok = await runCommand(snd[0], snd.slice(1));
@@ -217,7 +217,7 @@ export function createSystemNotifier(options: {
   return {
     /**
      * 系统通知（节流 + 超时杀进程；失败 resolve false 不 reject）。
-     * 1s 节流覆盖单次投递全部 spawn（toast + 自播同属一次逻辑操作，B3）；
+     * 1s 节流覆盖单次投递全部 spawn（toast + 自播同属一次逻辑操作）；
      * 节流吞掉时按上一次决议语义透传（返回「是否真有一次成功 spawn」）。
      * 命令构造走 text 域纯函数 buildSystemCommand/buildSoundCommand：
      * Windows：PowerShell WinRT toast / SoundPlayer 自播；macOS：osascript /
